requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-groq

# Add version constraints for stability

streamlit>=1.31.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain>=0.1.16
langchain-groq>=0.0.3
